services/crosspic/pyproject.toml
[project]
name = "crosspic"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.1",
    "pillow>=11.3.0",
]
